server/requirements.txt
pillow
Ipython
tensorflow
keras
keras_tqdm
numpy